x86: Add missing asmlinkages/asmregparm
Relies on the earlier patch to readd asmregparm.

This is an arch/x86 sweep to add asmlinkage or asmregparm
to all functions called by assembler code, including interrupt handlers,
syscalls and others.

I need this for a followon project that requires all functions
called by assembler to be marked in a special way.

The unmarked syscalls are marked asmregparm now. That's unusal
for syscalls -- normally they are all asmlinkage aka stack arguments.
I kept it at asmregparm to generate the same binary.

All calls with no argument are asmlinkage

All other functions that were unmarked before are asmregparm
now.

On 64bit it's all a nop currently.

I verified the patch generates the same binary as before.

I kept it all in a single patch for now. Please let me know
if you want it split up.

Cc: x86@kernel.org
Signed-off-by: Andi Kleen <ak@linux.intel.com>